% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstandard-package.r
\docType{package}
\name{rstandard}
\alias{rstandard}
\alias{rstandard-package}
\title{rstandard: This package contains function for statistics}
\description{
In statistics it is important calcualte standard deviation
}

