% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd.R
\name{standard_deviation}
\alias{standard_deviation}
\title{Calculate standard deviation}
\usage{
standard_deviation(x)
}
\arguments{
\item{x}{vector}
}
\value{
numeric
}
\description{
This function calculates standard deviation
}
\examples{

standard_deviation(c(1,2,3,4))

}

