% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{standard_error}
\alias{standard_error}
\title{Calculate standard error of the mean
This function calculates standard error}
\usage{
standard_error(x)
}
\arguments{
\item{x}{vector}
}
\value{
numeric
}
\description{
Calculate standard error of the mean
This function calculates standard error
}
\examples{

standard_error(c(1,2,3,4))

}

